Restructure synchronization.
